+++ Ass03Solution/BusinessObject/CartProduct.cs
@@ -1,44 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
 namespace BusinessObject
 {
-    public partial class CartProduct
+    internal class CartProduct
     {
-        public Product()
-        {
-            OrderDetails = new HashSet<OrderDetail>();
-        }
-
-        [Display(Name = "Product ID")]
-        public int ProductId { get; set; }
-
-        [Required]
-        [Display(Name = "Category")]
-        public int CategoryId { get; set; }
-
-        [Required]
-        [Display(Name = "Product Name")]
-        [StringLength(100, MinimumLength = 2)]
-        public string ProductName { get; set; }
-
-        [Required]
-        [Display(Name = "Weight")]
-        public string Weight { get; set; }
-
-        [Required]
-        [Display(Name = "Unit Price")]
-        public decimal UnitPrice { get; set; }
-
-        [Required]
-        [Display(Name = "Units In Stock")]
-        public int UnitsInStock { get; set; }
-
-        public virtual Category Category { get; set; }
-        public virtual ICollection<OrderDetail> OrderDetails { get; set; }
     }
 }
